Renamed generic variable names to be more descriptive, such as "fighter_index" and "attack_fighter_index"/"defend_fighter_index" in combat.c.

+++ include/combat.h
@@ -47,11 +47,11 @@
 /*! \name combat function prototypes  */
 int combat_check(int, int);     /*  entity.c  */
 int combat(int);                /*  intrface.c  */
-void battle_render(int, int, int);      /*  effects.c, [eh]skill.c, heroc.c, magic.c, selector.c  */
+void battle_render(signed int, size_t, int);      /*  effects.c, [eh]skill.c, heroc.c, magic.c, selector.c  */
 void draw_fighter(size_t, size_t);    /*  effects.c, hskill.c  */
-int fight(int, int, int);       /*  enemyc.c, [eh]skill.c, heroc.c  */
-void multi_fight(int);          /*  hskill.c  */
-void fkill(int);                /*  heroc.c, hskill.c, magic.c  */
+int fight(size_t, size_t, int); /*  enemyc.c, [eh]skill.c, heroc.c  */
+void multi_fight(size_t);       /*  hskill.c  */
+void fkill(size_t);             /*  heroc.c, hskill.c, magic.c  */
 
 
 /*!  global combat related variables  */
